+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for plugin manifest files with the "yml" extension. #3097
 - Randomize Agent binary hash (polymorphism) feature. #3244
 - Agent binary's SHA256 to `AgentRegistrationData`. #3244
+- `EmailAddress` identity type. #3270
 
 ### Changed
 - Renamed "Credential collector" to "Credentials collector". #3167